# apps/api/openapi.yaml
openapi: 3.0.0
paths:
...
  /word:
    get:
      operationId: WordController_getAll
      parameters:
        - name: offset
          required: true
          in: query
          schema:
            example: 0
            type: number
        - name: limit
          required: true
          in: query
          schema:
            example: 10
            type: number
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - processing
              - processed
        - name: id
          required: false
          in: query
          schema:
            type: number
        - name: created_at
          required: false
          in: query
          schema:
            format: date-time
            type: string
        - name: topic
          required: false
          in: query
          schema:
            type: number
        - name: word
          required: false
          in: query
          schema:
            type: string
        - name: catalog
          required: false
          in: query
          schema:
            type: number
        - name: language
          required: false
          in: query
          schema:
            type: string
            enum:
              - en
              - es
              - fr
              - de
              - it
              - ru
              - el
              - nl
        - name: audio
          required: false
          in: query
          schema:
            type: string
        - name: transcribtion
          required: false
          in: query
          schema:
            type: string
        - name: meaning
          required: false
          in: query
          schema:
            type: string
        - name: sortBy
          required: false
          in: query
          schema:
            example: created_at
            type: string
        - name: sortOrder
          required: false
          in: query
          schema:
            example: DESC
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetWordResponseDto'
        '400':
          description: Invalid param
        '500':
          description: Server error
      summary: Get all words
      tags: &ref_5
        - word
    put:
      operationId: WordController_update
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutWordRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PutWordResponseDto'
        '400':
          description: Invalid data
        '404':
          description: Word not found
        '500':
          description: Server error
      summary: Update word
      tags: *ref_5
...
  /word/generate:
    post:
      operationId: WordController_generateWords
      parameters:
        - name: language
          required: true
          in: query
          schema:
            type: string
        - name: topic
          required: true
          in: query
          schema:
            type: string
        - name: level
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Word generation started
      summary: Generate words
      tags: *ref_5
...
info:
  title: My API
  description: API description
  version: '1.0'
  contact: {}
...
servers: []
components:
  schemas:
...
    WordDto:
      type: object
      properties:
        status:
          type: string
          enum: &ref_8
            - processing
            - processed
        id:
          type: number
        created_at:
          type: string
          format: date-time
        topic:
          type: number
        word:
          type: string
        catalog:
          type: number
        language:
          type: string
          enum: &ref_9
            - en
            - es
            - fr
            - de
            - it
            - ru
            - el
            - nl
        audio:
          type: string
        transcribtion:
          type: string
        meaning:
          type: string
      required:
        - status
        - id
        - created_at
        - topic
        - word
        - catalog
        - language
        - audio
        - transcribtion
    GetWordResponseDto:
      type: object
      properties:
        total:
          type: number
          example: 100
        offset:
          type: number
          example: 0
        limit:
          type: number
          example: 10
        items:
          description: List of words
          type: array
          items:
            $ref: '#/components/schemas/WordDto'
      required:
        - total
        - offset
        - limit
        - items
    PutWordRequestDto:
      type: object
      properties:
        status:
          type: string
          enum: *ref_8
        id:
          type: number
        created_at:
          type: string
          format: date-time
        topic:
          type: number
        word:
          type: string
        catalog:
          type: number
        language:
          type: string
          enum: *ref_9
        audio:
          type: string
        transcribtion:
          type: string
        meaning:
          type: string
    PutWordResponseDto:
      type: object
      properties:
        status:
          type: string
          enum: *ref_8
        id:
          type: number
        created_at:
          type: string
          format: date-time
        topic:
          type: number
        word:
          type: string
        catalog:
          type: number
        language:
          type: string
          enum: *ref_9
        audio:
          type: string
        transcribtion:
          type: string
        meaning:
          type: string
      required:
        - status
        - id
        - created_at
        - topic
        - word
        - catalog
        - language
        - audio
        - transcribtion
...